Try using Content-Type charset if the file is not UTF-8
If the po file encoding is not UTF-8, try the charset specified
in the Content-Type header field, if it all fails, display a
message suggesting msgconv/iconv to convert the file by hand.

Once a file is converted successfully to UTF-8, its Content-Type
field will be changed, and subsequent saves will happen in UTF-8
encoding.

The glib dependency has been bumped to 2.32 because of
g_file_new_tmp().

https://bugzilla.gnome.org/show_bug.cgi?id=566358