Fix assistant
The assistant wasn't working because there was two pages with the same
title, and that was causing problems. I've renamed the languages fetcher
page to "Language".

I've also set this assistant as a modal window attached to the main
window to show it centered and with the main window disabled and dark.