Functional app with *modern* interface
This patch tries to leave the app functional so it can be used meantime
we continue working in the UI.

First a *Project* selection is showed, showing recent files so you can
open files from there. This view will be updated soon to a
GtrProjectSelector widget, that will manage more information than only a
.po file, but for now, a *project* is a single po file.

Once a file is opened the NoteBook is showed using the old interface.
The toolbar and status bar are removed so we need to show that
information in other places, following the gnome HIG.

Menu actions are moved to the headerbar or to the hamburger menu.