Don't compare msgid on selection change
Comparing the msgid on selection change is causing problems because
it's possible to have multiple msgs with the same msgid and different
context.

We could compare also the context, but I think that there's not much
waste in send the signal and do all the work if it's the same message,
the check complicate the code and don't give us an appreciable
performance improvement, so I've removed it.

Fix #19