Find and replace to the main menu
I've added two new entries to the main menu to show the find dialog and
the find and replace dialog.

In the future we should remove these dialogs and use a GtkRevealer in
the top with the search info and controls and maybe we can unify these
two dialogs in one with an advanced button.

Fix #23