message-table: Show msg status inline
I've removed the STATUS and the ID columns and to show the message
status now we use text colors so fuzzy will be orange and untranslated
red. These two colors are getted from the theme so this should work in
the dark mode and other themes should work correctly too.

Fix #25