List of languages in a config file instead iso-codes
The iso codes has the list of languages codes and the list of country
codes, but there's no a list of these two mixed.

To have a custom list with language codes with region we can provide a
custom list, like we provide the plural forms.

In the future this will help us to translate language names too. Right
now we've a list of languages in plain english.

See #43