Use theme_fg_color for translated texts
We were using "fg_color" for translated text and that was not working
correctly in the dark theme so I changed to the "theme_fg_color" and it
works correctly now.

I've also moved the color check from the theme from the widget init to
the cell draw to make the color change dynamic.

Fix #46