Show search dialog information in dialog
The find and replace dialog was using the statusbar to show messages
like the "not found", we're not using the statubar anymore so we need to
find a new way to show this kind of information to the user.

This patch adds a new widget to the search dialog with a revealer and a
label that will be used to show erros and other information in this
dialog.

Fix #73