Add Generic filter selection widget
We've a custom widget for the language selection with a long text list
and a filter text entry. We can use this widget in the DL selection
because we're currenlty using comboboxes and we've a lot of options,
it'll be easier if we reuse this widget.

This patch generalizes the widget to be able to use there.