Add the Open from Damned Lies(DL) button
A user can view two buttons in the main window when no file is opened.
These buttons are namely: Open & Open from DL.

However, if a file is opened, the user can view only the Open button.
Thus, a user cannot load files from Damned Lies.

To fix this make the Open from DL button visible when a file is opened.
Prompt a user to save an edited file before the open button is opened.

Closes #65