codeview: Don't link to not found files
The source path inside po files is relative to the project root. If you
open the po file placed inside a gnome project this should work, but you
can open a po file placed anywhere so instead of show a broken link,
it's better to show the path as plain text.

This patch tries to find the source file in ../PATH relative to the .po
location.

A new function called find_source_file has been added and can be
improved to try to find the source file.

See https://gitlab.gnome.org/GNOME/gtranslator/issues/100