Use base language to get the plural form
This patch removes some duplicate strings in the plural forms document
and also provides a fallback in the case of a language is not found to
use the base language. For example if "en_US" is not found it'll try to
use the "en" entry in the document.

Fix https://gitlab.gnome.org/GNOME/gtranslator/-/issues/109