Download DL files to Download folder
We shouldn't use the /tmp folder directly because if the app is
flatpaked the /tmp folder is not accessible by the host.

This patch changes the default DL file location to the Download folder,
so it's not possible to loose the work when the user closes the app
without changing the file location. By default downloaded po files will
be stored on $HOME/Downloads.

With this new behavior we don't need to ask to the user for a new
location just after the file is downloaded. We just stored the file in
Downloads by default and the user can do a normal app usage.

Fix https://gitlab.gnome.org/GNOME/gtranslator/-/issues/117