Font change in preferences
Instead of change just the font size with three buttons it's better to
provide a way to change the font and font size in the preferences
editor.

The editor font is a preference so it's better placed there. This patch
provides a font-chooser button so now it's possible to change the font
and font size at once in the preferences dialog.

See https://gitlab.gnome.org/GNOME/gtranslator/-/issues/120