Use GtkFileChooserNative
This patch uses the GtkFileChooserNative instead of the GtkFileChooser
to open and save po files. This will improve the experience for flatpak
because it will show bookmarks and the integration with the host system
is better.

This change also introduces a small issue with file paths. The
GtkFileChooserNative returns a path like `/run/user/...` for each file
so it's not possible to know the real path of that file.

We were showing the full file path on the headerbar, but with this temp
path that's not useful, so this patch removes that subtitle.

You can notice this problem also when clicking on "save as" because the
default folder will be that temp path, instead the user folder.

Fix https://gitlab.gnome.org/GNOME/gtranslator/-/issues/116