gtrl-dl-teams, gtr-application: Add libhandy widgets
When you click on the "load from DL" button with the window in
full screen, the form takes the whole width. It will be good if the
form has a maximum width.

Use libhandy to extend Gtk so as to make the form smoothly grow to a
maximum width and automatically adapt its size to that of it's parent.

Closes: https://gitlab.gnome.org/GNOME/gtranslator/-/issues/122