gtr-dl-teams: Reserve a module for translation
It is currently possible to reserve a module for translation from the
Damned Lies website. Enabling translators to perform the same action
directly from the Translation Editor will be a very useful feature.

- Enable the reserve button for any of the valid states as illustrated
on the vertimus workflow https://l10n.gnome.org/help/vertimus_workflow/
and disable otherwise.
- Perform a POST operation on the Damned Lies API (using the reserve
endpoint) when the reserve button is clicked.