Do not override files on DL load
Instead of override the files when downloading from DL it's better to
just add a number to the filename and create a new one so we do not
override the existing files on the user Downloads folder.

This patch adds a number to the filename so we'll have something like
"es_ES (1).po".

https://gitlab.gnome.org/GNOME/gtranslator/-/issues/124