gtr-profile-dialog: Add a new field for auth token
Users need to be authenticated in order to perform a reserve for
translation operation or to upload a file. So there is a need to make
room for authentication.

Add a new field to the profile dialog to store a user's
authentication token and save the it

Add a tooltip to give more information about the authentication token