Disable redirection on DL requests
This patch disables the redirecction management in the soup message sent
to the damned lies server to receive the real response 302 status code
when the authentication code is not correct.

This will fix the problem of not having configured the token and trying
to reserve a module with a better error message.

Fix https://gitlab.gnome.org/GNOME/gtranslator/-/issues/129