Refactoring of GtrContext panel
Replace the TextView widget with a series of widgets to show in the
context panel, the translator comments, the source comments and file
paths.

Paths are not clickable right now, that's something that we should
(re)implement in the future, it's not too useful when the file is not
opened from a git clone so it should be improved.