Set run_command check to false
Since Meson 0.61.0, run_command emits a warning when used
without 'check' set with a Boolean. That's because its
default value is currently 'false' but will change to 'true'.

This commit sets to true in order to NOT skip error.

Message:

WARNING: You should add the boolean check kwarg to the run_command call.
         It currently defaults to false,
         but it will default to true in future releases of meson.
         See also: https://github.com/mesonbuild/meson/issues/9300

Link to note:
https://mesonbuild.com/Release-notes-for-0-61-0.html#warning-if-check-kwarg-of-run_command-is-missing