Do not unref file objects on application_setup_window
The file list in the gtr_application_setup_window is owned by the caller
so if we dereference here it causes a crash. This patch just frees the
memory used by the list but not the file objects in the list that will
be freed later.

Fix https://gitlab.gnome.org/GNOME/gtranslator/-/issues/182