Remove setting to hide flap
There is no in-app way to show the panel again if the setting is set to
false. some things like hiding the panel using the inspector, or very
old configs, could set the setting to `False` locking out the user from
using the panel.

I don't think this is fit for the Preferences window, a better fit might
be a toggle in the header bar. But further work is required to remove
other buttons and make room for it. Then the setting can be
reintroduced.