Reimplement the search
This patch reimplement the search doing a simple for-loop with a regex
to search in the list of messages and updating the interface just when
the message is found.

A following step could be to do this async, in a different thread, so it
won't freeze the interface for large files, but for now this solution
should be good enough, and should improve the search a lot.

Fix https://gitlab.gnome.org/GNOME/gtranslator/-/issues/194
See https://gitlab.gnome.org/GNOME/gtranslator/-/issues/144