Fix back button on Load from DL
This patch fixes the issue of clicking back and load from DL again
without a file opened. When the page "dlteams" exists but the page
"poeditor" doesn't it's not possible to call replace_with_tags with
"poeditor", so we need to use the current page in the navigation_view.

In the future we can use the adw_navigation_view_get_visible_page_tag
directly instead of getting the AdwNavigationPage and calling get_tag,
but this function is still unstable and will be available since version
1.7.

https://gnome.pages.gitlab.gnome.org/libadwaita/doc/main/method.NavigationView.get_visible_page_tag.html