msg: Remove leftover plurals on fix
This patch makes sure to remove all not needed plural msgstr from the
po file when fixing plurals.

The API `po_message_set_msgstr_plural` reduces the number of plurals
passing NULL with the "last" index. It only works with the last one so
we need to do in reverse to make sure that we're removing all possible
exceding messages.

So for example if plurals is 1 and we've 4 plurals for a message in this
broken file we need to:

 1. Get the index for the last plural translation (3 in this example)
 2. Call po_message_set_msgstr_plural (msg, 3, NULL)
 3. Call po_message_set_msgstr_plural (msg, 2, NULL)
 4. Call po_message_set_msgstr_plural (msg, 1, NULL)

So it only exists msgstr[0] in the file at the end.