Fix close confirmation window on ctrl+w
With the movement from the signal to the virtual method, the window is
closed always, even before the close confirmation dialog is present.

This patch "return TRUE" in the close_request virtual method, so the
window is not closed here. The function gtr_file_quit will close the
window after the save state is checked and the confirmation dialog is
closed.