Move remove_tab to gtr_window_set_po before creation
We shouldn't call to gtr_window_remove_tab before trying to load,
because if the loading fails, the application is in an inconsistent
state.

This patch moves the call to remove_tab just before the new tab creation
so there's no a change to leave the window without a working tab.